pyproject.toml
[project]
name = "symularaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.15",
    "langchain-text-splitters>=1.1.2",
    "langchain[openai]>=1.2.15",
    "langgraph>=1.1.8",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sse-starlette>=3.3.4",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.44.0",
]
